Don't push as part of the workflow
Pushing automatically at the end of the build felt very awkard and
caused problems with pushing the build log, which I'd like to have in
the package but is being actively written during a push target. The most
natural way to push is just manually running the push script when it's
time to push.